backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
motor>=3.3.2
pymongo>=4.6.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Stripe Payments
stripe>=7.12.0

# Google Gemini (standalone - minimal dependencies)
google-generativeai>=0.3.2

# Cloud Storage
boto3>=1.34.25

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
Pillow>=11.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.3
